pyproject.toml
[tool.poetry]
name = "whistling"
version = "0.0.0"
description = "A collection of functions to report a malicious URL"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
httpx = ">=0.24,<1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3"
environs = "^9.5"
mypy = "^1.3"
pre-commit = "^3.3"
pytest = "^7.3"
pytest-cov = "^4.0"
pytest-randomly = "^3.12"
pytest-sugar = "^0.9"
requests = "2.29.0"
vcrpy = "^4.2"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
